src/tests/requirements.txt
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-timeout==2.2.0
httpx==0.25.2
asyncpg==0.29.0
freezegun==1.4.0
pyyaml>=6.0
fastapi==0.104.1
requests==2.31.0
flake8==7.0.0
black==23.12.1
